requirements.txt
# Data
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
pyarrow>=14.0.0

# Synthetic Data
faker>=19.0.0
sdv>=1.9.0

# ML / Clustering
scikit-learn>=1.3.0
umap-learn>=0.5.5
hdbscan>=0.8.33
matplotlib>=3.7.0
seaborn>=0.12.0

# GCP
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
google-auth>=2.22.0

# Gemini
google-generativeai>=0.7.0

# Utils
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0

# Tooling
ruff>=0.1.0
mypy>=1.5.0